gio-2.0.pc: Avoid full path to executables
Instead of using full path in glib_compile_schemas and
glib_compile_resources variables, rely on having the executables in
PATH.

This fixes the cross-compiling case where we cannot execute the binaries
for target system. Instead of executing the target system's binaries, we
need to use the native versions installed on the build host. The easiest
way to find the native executables is to just pick them up from PATH.

In addition, this brings gio-2.0.pc in line with glib-2.0.pc -- the
latter has historically only listed the executable name and not the full
path.

https://bugzilla.gnome.org/show_bug.cgi?id=673911