gmain: Optimize insertion of matching or higher priority sources
Calling g_source_attach() does an O(n) iteration over all current
sources until it found the insertion point, while holding the context
mutex.  This is a serious performance hit for heavy asynchronous I/O,
because GIOScheduler sends results back to the source context via
g_source_attach().

A more correct fix for this would be a priority queue, but this patch
is quite simple - we just walk the source list in reverse instead of
forwards.  For applications where most sources are the same priority
(or they have just a few lower-priority sources), this patch is a
large speedup.

However, this patch will degrade performance in the case of inserting
a high-priority source for applications which have many lower-priority
sources.

https://bugzilla.gnome.org/show_bug.cgi?id=619329